# api/public/swagger.json
{
	"components": {
...
		"schemas": {
			"LoginUserResponse": {
				"properties": {
					"email": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetAdminProfileResponse": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MessageResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Location": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"seat": {
						"type": "number",
						"format": "double"
					},
					"shows": {
						"items": {
							"$ref": "#/components/schemas/Show"
						},
						"type": "array"
					},
					"staffs": {
						"items": {
							"$ref": "#/components/schemas/Staff"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"address",
					"phone",
					"seat",
					"shows",
					"staffs"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Show": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"startDate": {
						"type": "string"
					},
					"startSlot": {
						"type": "number",
						"format": "double"
					},
					"isActive": {
						"type": "boolean"
					},
					"location": {
						"$ref": "#/components/schemas/Location"
					},
					"locationId": {
						"type": "number",
						"format": "double"
					},
					"movie": {
						"$ref": "#/components/schemas/Movie"
					},
					"movieId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"startDate",
					"startSlot",
					"isActive",
					"location",
					"locationId",
					"movie",
					"movieId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Movie": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"length": {
						"type": "number",
						"format": "double"
					},
					"isActive": {
						"type": "boolean"
					},
					"shows": {
						"items": {
							"$ref": "#/components/schemas/Show"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"length",
					"isActive",
					"shows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Staff": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"location": {
						"$ref": "#/components/schemas/Location"
					},
					"locationId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"password",
					"isActive",
					"location",
					"locationId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GetCustomerProfileResponse": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GetStaffProfileResponse": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "movie-ticket-online-booking-system",
		"version": "1.0.0",
		"description": "It is the project of Service Oriented Application course.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/admin/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Login admin",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/admin/profile": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Get admin profile",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetAdminProfileResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/admin/staff": {
			"post": {
				"operationId": "CreateStaff",
				"responses": {
					"201": {
						"description": "Created staff",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"locationId": {
										"type": "number",
										"format": "double"
									},
									"name": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"locationId",
									"name",
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetStaffes",
				"responses": {
					"200": {
						"description": "Get all staff",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Staff"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/admin/staff/{id}": {
			"get": {
				"operationId": "GetStaff",
				"responses": {
					"200": {
						"description": "Get staff",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Staff"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateStaff",
				"responses": {
					"201": {
						"description": "Update staff",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteStaff",
				"responses": {
					"201": {
						"description": "Delete staff",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/location": {
			"post": {
				"operationId": "CreateLocation",
				"responses": {
					"201": {
						"description": "Created location",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"seat": {
										"type": "number",
										"format": "double"
									},
									"phone": {
										"type": "string"
									},
									"address": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"seat",
									"phone",
									"address",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/admin/location/{id}": {
			"patch": {
				"operationId": "UpdateLocation",
				"responses": {
					"201": {
						"description": "Update location",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"phone": {
										"type": "string"
									},
									"address": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/admin/movie": {
			"post": {
				"operationId": "CreateMovie",
				"responses": {
					"201": {
						"description": "Created movie",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"length": {
										"type": "number",
										"format": "double"
									},
									"description": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"length",
									"description",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/admin/movie/{id}": {
			"patch": {
				"operationId": "UpdateMovie",
				"responses": {
					"201": {
						"description": "Update movie",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"description": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/customer/profile": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Get customer profile",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetCustomerProfileResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Customer"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
		"/staff/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Login admin",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Staff"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/staff/profile": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Get staff profile",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetStaffProfileResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Staff"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/api",
			"description": "Local development"
		}
	]
}